[libvelocity_control] Fix an issue in pwm_cytron
When the cytron was going from 100 or -100 to 0 there was an issue.
Also improve the pwm to be type consistent and more efficient.

Remove the TODO about base clock because it is read from a particular
memory space so it would prevent cross-compilation. There may be
another solution though.

Add some functions to tests the equality between to float numbers.